test: add database test
