+++ collab-database/tests/database_test/field_test.rs
@@ -1,7 +1,6 @@
+use crate::database_test::helper::{create_database, create_database_with_default_data};
 use collab_database::fields::Field;
 use collab_database::views::CreateViewParams;
-
-use crate::helper::{create_database, create_database_with_default_data};
 
 #[test]
 fn create_single_field_test() {
