+++ collab-database/tests/database_test/row_test.rs
@@ -1,8 +1,7 @@
+use crate::database_test::helper::{create_database, create_database_with_default_data};
 use collab_database::block::CreateRowParams;
 use collab_database::database::gen_row_id;
 use collab_database::views::CreateViewParams;
-
-use crate::helper::{create_database, create_database_with_default_data};
 
 #[test]
 fn create_row_shared_by_two_view_test() {
