+++ collab-database/tests/database_test/view_test.rs
@@ -7,10 +7,10 @@ use collab_database::views::{
 };
 use nanoid::nanoid;
 use serde_json::json;
-
 use assert_json_diff::assert_json_eq;
+use crate::database_test::helper::{create_database, create_database_with_default_data};
+use crate::helper::TestFilter;
 
-use crate::helper::{create_database, create_database_with_default_data, TestFilter};
 
 #[test]
 fn create_initial_database_test() {
