+++ collab-database/tests/main.rs
@@ -1,0 +1,3 @@
+mod database_test;
+mod helper;
+mod user_test;
