+++ collab-database/tests/user_test/type_option_test.rs
@@ -1,8 +1,8 @@
 use collab::core::any_map::AnyMapExtension;
 use collab_database::fields::{Field, TypeOptionDataBuilder, TypeOptions};
 use collab_database::views::CreateDatabaseParams;
+use crate::user_test::helper::{user_database_test, UserDatabaseTest};
 
-use crate::helper::{user_database_test, UserDatabaseTest};
 
 #[test]
 fn update_single_type_option_data_test() {
