+++ collab-document/tests/delta/util.rs
@@ -8,7 +8,7 @@ use nanoid::nanoid;
 use serde_json::{json, Value};
 use std::collections::HashMap;
 use std::ops::Deref;
-use std::path::PathBuf;
+
 use std::rc::Rc;
 use std::sync::{Arc, Once};
 use tempfile::TempDir;
